+++ client/src/cmds-internal/index.ts
@@ -1,7 +1,9 @@
 import { OpenDebugFile } from './OpenDebugFile';
+import { OpenDebugFileSharp } from './sharp/OpenDebugFile';
 
 export const CommandsInternal = {
     allCommandes: {
         [OpenDebugFile.cmd]: OpenDebugFile,
+        [OpenDebugFileSharp.cmd]: OpenDebugFileSharp,
     },
 }