+++ server/src/language-services/json/definition.ts
@@ -62,3 +62,43 @@ export interface AventusConfig {
 		[alias: string]: string
 	}
 }
+
+export interface AventusSharp {
+	csProj: string,
+	outputPath: string,
+	exportEnumByDefault?: boolean,
+	exportStorableByDefault?: boolean,
+	exportHttpRouteByDefault?: boolean,
+	exportErrorsByDefault?: boolean,
+	exportWsEndPointByDefault?: boolean,
+	exportWsEventByDefault?: boolean,
+	exportWsRouteByDefault?: boolean,
+	replacer?: {
+		all?: AventusSharpReplacerPart
+		genericError?: AventusSharpReplacerPart
+		httpRouter?: AventusSharpReplacerPart
+		normalClass?: AventusSharpReplacerPart
+		storable?: AventusSharpReplacerPart
+		withError?: AventusSharpReplacerPart
+		wsEndPoint?: AventusSharpReplacerPart
+		wsEvent?: AventusSharpReplacerPart
+		wsRouter?: AventusSharpReplacerPart
+	},
+}
+
+interface AventusSharpReplacerPart {
+	type: {
+		[key: string]: {
+			result: string,
+			file?: string,
+			useTypeImport?: boolean,
+		}
+	}
+	result: {
+		[key: string]: {
+			result: string,
+			file?: string,
+			useTypeImport?: boolean,
+		}
+	}
+}