+++ server/src/language-services/ts/parser/ParserTs.ts
@@ -141,6 +141,7 @@ export class ParserTs {
     public functions: { [shortName: string]: FunctionInfo } = {};
     public enums: { [shortName: string]: EnumInfo } = {};
     public imports: { [importClassName: string]: ImportInfo } = {};
+    public packages: { [importClassName: string]: { fullname: string } } = {};
     public npmImports: {
         [importClassName: string]: {
             nameInsideLib: string,